+++ services/QuillLMS/client/app/bundles/Shared/libs/tests/fillInBlankInputStyle.test.tsx
@@ -1,4 +1,12 @@
-import { fillInBlankInputStyle, determineShortestCue, generateSpan } from '../fillInBlankInputStyle';
+import { fillInBlankInputStyle, } from '../fillInBlankInputStyle';
+import * as helpers from '../fillInBlankInputStyleHelpers';
+
+jest.mock('../fillInBlankInputStyleHelpers', () => ({
+  ...jest.requireActual('../fillInBlankInputStyleHelpers'),
+  generateOffsetWidth: jest.fn(),
+}));
+
+const { determineShortestCue, } = helpers
 
 describe('determineShortestCue function', () => {
   it('returns the shortest cue from a list of cues', () => {
@@ -23,9 +31,14 @@ describe('determineShortestCue function', () => {
 });
 
 describe('fillInBlankInputStyle function', () => {
+  beforeEach(() => {
+    helpers.generateOffsetWidth.mockImplementation((element) => {
+      return element.textContent.length * 10;
+    });
+  });
 
   it('calculates width when there is a value and cues', () => {
-    const { width } = fillInBlankInputStyle('aaaa', ['a', 'longercue', 'muchlongercue'], true);
+    const { width } = fillInBlankInputStyle('aaaa', ['a', 'longercue', 'muchlongercue']);
     expect(width).toBe('60px'); // Mocked 40 for 'aaaa' + 20 padding
   });
 
@@ -35,18 +48,18 @@ describe('fillInBlankInputStyle function', () => {
   });
 
   it('calculates width when value and cues are null', () => {
-    const { width } = fillInBlankInputStyle(null, null, true);
+    const { width } = fillInBlankInputStyle(null, null);
     expect(width).toBe('20px'); // Mocked 0 for null + 20 padding
   });
 
   it('assigns the textAlign value to center and paddingRight to 10px when the value is shorter than the shortest cue', () => {
-    const { textAlign, paddingRight, } = fillInBlankInputStyle('a', ['aaaa', 'longercue', 'muchlongercue'], true);
+    const { textAlign, paddingRight, } = fillInBlankInputStyle('a', ['aaaa', 'longercue', 'muchlongercue']);
     expect(textAlign).toBe('center');
     expect(paddingRight).toBe("10px");
   });
 
   it('assigns the textAlign value to left and paddingRight to 0px when the valueSpan is longer than the shortest cue', () => {
-    const { textAlign, paddingRight } = fillInBlankInputStyle('aaaa', ['a', 'longercue', 'muchlongercue'], true);
+    const { textAlign, paddingRight } = fillInBlankInputStyle('aaaa', ['a', 'longercue', 'muchlongercue']);
     expect(textAlign).toBe('left');
     expect(paddingRight).toBe("0px");
   });
